extract county selector, add local storage hook
