+++ package.json
@@ -3,10 +3,12 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
+    "@react-hookz/web": "^12.3.0",
     "autoprefixer": "^10.4.2",
     "date-fns": "^2.28.0",
     "papaparse": "^5.3.1",
     "postcss": "^8.4.5",
+    "ramda": "^0.28.0",
     "react": "^17.0.2",
     "react-dom": "^17.0.2",
     "react-query": "^3.34.8",
